refactor: move list under collection
