--- kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/list/ListAliases.kt
@@ -1,21 +1,0 @@
-package xyz.lbres.kotlinutils.list
-
-/**
- * List of Booleans
- */
-typealias BoolList = List<Boolean>
-
-/**
- * List of Ints
- */
-typealias IntList = List<Int>
-
-/**
- * List of Strings
- */
-typealias StringList = List<String>
-
-/**
- * List of values paired with weights
- */
-typealias WeightedList<T> = List<Pair<T, Float>>
