+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/random/RandomExt.kt
@@ -1,6 +1,6 @@
 package xyz.lbres.kotlinutils.random
 
-import xyz.lbres.kotlinutils.list.WeightedList
+import xyz.lbres.kotlinutils.collection.list.WeightedList
 import kotlin.random.Random
 
 /**
