+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/collection/IntCollectionExtTest.kt
@@ -1,6 +1,6 @@
 package xyz.lbres.kotlinutils.collection
 
-import xyz.lbres.kotlinutils.list.IntList
+import xyz.lbres.kotlinutils.collection.list.IntList
 import kotlin.test.Test
 import kotlin.test.assertEquals
