+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/collection/list/MutableListExtTest.kt
@@ -1,8 +1,8 @@
-package xyz.lbres.kotlinutils.list
+package xyz.lbres.kotlinutils.collection.list
 
-import xyz.lbres.kotlinutils.list.mapInPlace
-import xyz.lbres.kotlinutils.list.mapInPlaceIndexed
-import xyz.lbres.kotlinutils.list.setAllValues
+import xyz.lbres.kotlinutils.collection.list.mapInPlace
+import xyz.lbres.kotlinutils.collection.list.mapInPlaceIndexed
+import xyz.lbres.kotlinutils.collection.list.setAllValues
 import xyz.lbres.kotlinutils.utils.simpleIf
 import kotlin.test.Test
 import kotlin.test.assertEquals
