+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/collection/multiset/const/ConstMutableMultiSetTest.kt
@@ -7,7 +7,7 @@ import xyz.lbres.kotlinutils.collection.multiset.impl.ConstMutableMultiSetImpl
 import xyz.lbres.kotlinutils.collection.multiset.impl.MultiSetImpl
 import xyz.lbres.kotlinutils.collection.multiset.testutils.* // ktlint-disable no-wildcard-imports no-unused-imports
 import xyz.lbres.kotlinutils.internal.constants.Suppressions
-import xyz.lbres.kotlinutils.list.IntList
+import xyz.lbres.kotlinutils.collection.list.IntList
 import kotlin.test.Test
 
 class ConstMutableMultiSetTest {
