+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/collection/multiset/impl/mutateTests.kt
@@ -3,7 +3,7 @@ package xyz.lbres.kotlinutils.collection.multiset.impl
 import xyz.lbres.kotlinutils.collection.multiset.MutableMultiSet
 import xyz.lbres.kotlinutils.collection.multiset.mutableMultiSetOf
 import xyz.lbres.kotlinutils.collection.multiset.testutils.*
-import xyz.lbres.kotlinutils.list.StringList
+import xyz.lbres.kotlinutils.collection.list.StringList
 import kotlin.test.assertEquals
 
 fun runMutableElementsAddTests() {
