+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/collection/multiset/testutils/commonFilterToSetTests.kt
@@ -4,7 +4,7 @@ import xyz.lbres.kotlinutils.collection.multiset.ConstMultiSet
 import xyz.lbres.kotlinutils.collection.multiset.MultiSet
 import xyz.lbres.kotlinutils.collection.multiset.emptyMultiSet
 import xyz.lbres.kotlinutils.collection.multiset.multiSetOf
-import xyz.lbres.kotlinutils.list.IntList
+import xyz.lbres.kotlinutils.collection.list.IntList
 import kotlin.test.assertContains
 import kotlin.test.assertEquals
