+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/collection/multiset/testutils/commonMapToSetTests.kt
@@ -4,8 +4,8 @@ import xyz.lbres.kotlinutils.collection.multiset.ConstMultiSet
 import xyz.lbres.kotlinutils.collection.multiset.MultiSet
 import xyz.lbres.kotlinutils.collection.multiset.emptyMultiSet
 import xyz.lbres.kotlinutils.collection.multiset.multiSetOf
-import xyz.lbres.kotlinutils.list.IntList
-import xyz.lbres.kotlinutils.list.copyWithoutLast
+import xyz.lbres.kotlinutils.collection.list.IntList
+import xyz.lbres.kotlinutils.collection.list.copyWithoutLast
 import xyz.lbres.kotlinutils.utils.simpleIf
 import kotlin.test.assertContains
 import kotlin.test.assertEquals
