+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/collection/multiset/testutils/commonMutateTests.kt
@@ -3,7 +3,7 @@ package xyz.lbres.kotlinutils.collection.multiset.testutils
 import xyz.lbres.kotlinutils.collection.multiset.MutableMultiSet
 import xyz.lbres.kotlinutils.collection.multiset.emptyMultiSet
 import xyz.lbres.kotlinutils.collection.multiset.multiSetOf
-import xyz.lbres.kotlinutils.list.StringList
+import xyz.lbres.kotlinutils.collection.list.StringList
 import kotlin.test.assertEquals
 
 fun runAddTests(createMutableSet: (Collection<*>) -> MutableMultiSet<*>) {
