+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/random/RandomExtTest.kt
@@ -1,6 +1,6 @@
 package xyz.lbres.kotlinutils.random
 
-import xyz.lbres.kotlinutils.list.WeightedList
+import xyz.lbres.kotlinutils.collection.list.WeightedList
 import xyz.lbres.kotlinutils.runTestWithWeights
 import kotlin.random.Random
 import kotlin.test.Test
